TEMP CHANGES

+++ pyscripts/run.py
@@ -5,11 +5,11 @@ import os
 
 experiments_path = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'experiments'   ))
 grief_path 		 = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'grief_history' )) 
-algorithms 		 = ['rand_1_bin', 'rand_to_best_1_bin']
-# algorithms 		 = ['best_1_bin']
+# algorithms 		 = ['rand_1_bin', 'rand_to_best_1_bin']
+algorithms 		 = ['rand_1_bin']
 configs = ConfigParser()
 
-ng = 3
+ng = 10
 
 for algorithm in algorithms:
 
@@ -71,8 +71,9 @@ for algorithm in algorithms:
 			cmd = 'cp opposite_generations ' + os.path.join(execution_path, 'opposite_generations')
 			os.system(cmd)
 			
-			cmd = "rm opposite_generations && touch opposite_generations" 
-			os.system(cmd)
+			if obl:
+				cmd = "rm opposite_generations && touch opposite_generations" 
+				os.system(cmd)
 
 			cmd = "rm -rf grief_history/*" 
 			os.system(cmd)
